+++ .eslintrc.js
@@ -22,6 +22,6 @@ module.exports = {
   },
   ignorePatterns: [
     '/libs/deps/*',
-    '/tools/loc/*',
+    '/tools/*',
   ],
 };
