--- tools/bulk-caas/index.html
@@ -1,10 +1,0 @@
-<html>
-  <body>
-    <a id="save" href="#">Save</a>
-    <script type="module">
-      import init from './bulk-caas.js';
-      const saveLink = document.querySelector('#save');
-      saveLink.addEventListener('click', () => init());
-    </script>
-  </body>
-</html>
